consolidated pointcutlatch into hapi-fhir-jpaserver-model
